Module Pbkdf.Make

Parameters

module H : Mirage_crypto.Hash.S

Signature

val pbkdf1 : password:Cstruct.t -> salt:Cstruct.t -> count:int -> dk_len:int -> Cstruct.t

pbkdf1 password salt count dk_len is dk, the derived key of dk_len octets. The salt must be eight octets, count the iteration count.

raises Invalid_argument

when either salt is not eight octets long or either count or dk_len are not valid.

val pbkdf2 : password:Cstruct.t -> salt:Cstruct.t -> count:int -> dk_len:int32 -> Cstruct.t

pbkdf2 password salt count dk_len is dk, the derived key of dk_len octets.

raises Invalid_argument

when either count or dk_len are not valid