Module Pbkdf

module type S = sig ... end

RFC 2898 specifies two password-based key derivation functions (PBKDF1 and PBKDF2), which are abstracted over a specific hash/pseudorandom function.

module Make : functor (H : Mirage_crypto.Hash.S) -> S

Given a Hash/pseudorandom function, get the PBKDF

val pbkdf1 : hash:[ `MD5 | `SHA1 ] -> password:Cstruct.t -> salt:Cstruct.t -> count:int -> dk_len:int -> Cstruct.t

convenience pbkdf1 hash password salt count dk_len where the hash has to be provided explicitly

val pbkdf2 : prf:Mirage_crypto.Hash.hash -> password:Cstruct.t -> salt:Cstruct.t -> count:int -> dk_len:int32 -> Cstruct.t

convenience pbkdf2 prf password salt count dk_len where the prf has to be provided explicitly