Module Pbkdf

module type S = sig ... end

RFC 2898 specifies two password-based key derivation functions (PBKDF1 and PBKDF2), which are abstracted over a specific hash/pseudorandom function.

module Make (H : Digestif.S) : S

Given a Hash/pseudorandom function, get the PBKDF

val pbkdf1 : hash:Digestif.hash' -> password:string -> salt:string -> count:int -> dk_len:int -> string

convenience pbkdf1 hash password salt count dk_len where the hash has to be provided explicitly

val pbkdf2 : prf:Digestif.hash' -> password:string -> salt:string -> count:int -> dk_len:int32 -> string

convenience pbkdf2 prf password salt count dk_len where the prf has to be provided explicitly